chore(core-manager): rename authentication classes to validator
