+++ packages/core-manager/src/ioc/identifiers.ts
@@ -3,6 +3,6 @@ export const Identifiers = {
     HTTPS: Symbol.for("API<HTTPS>"),
     ActionReader: Symbol.for("Discover<Action>"),
     PluginFactory: Symbol.for("Factory<Plugin>"),
-    BasicAuthentication: Symbol.for("Authentication<Basic>"),
-    TokenAuthentication: Symbol.for("Authentication<Token>"),
+    BasicCredentialsValidator: Symbol.for("Validator<BasicCreadentials>"),
+    TokenValidator: Symbol.for("Validator<Token>"),
 };
